device-connectors/src/testflinger_device_connectors/devices/iotscript/iot-auto-sanity/pyproject.toml
[project]
name = "auto-sanity"
description = "sanity tool for IoT project"
readme = "README"
dependencies = [
    "jsonschema",
    "pyyaml",
    "pyserial",
    "schedule",
    "wrapt_timeout_decorator",
]

dynamic = ["version"]

[project.scripts]
auto-sanity = "sanity.agent_entry:main"
